I did not see a README in this directory so I just included my reasons for adding these papers in the commit log.
The dynamic window approach to collision avoidance is an influential
paper for mobile robots. The method is based on a robot's dynamics
rather than higher-level representations of a robot and/or obstacles in
an environment.

The PRM and RRT algorithms are two seminal papers in robot motion
planning. The problem of motion planning scales exponentially with the
degrees of freedom a robot has and the degrees of freedom the obstacles
in an environment have. Thus, planning with high degrees of freedom leads to many problems
such as incompleteness and extremely slow speed. The PRM method was the first to
propose a sampling-based stratey to deal with motion planning and
created a practical methed for offline planning of robot manipulators.
The RRT method modified PRM by using a tree structure rather than a
graph so that non-holonomic and other constraints could be considered
when planning.

The Instantaneous Trajectory Generation method is relatively new, but
very important. It allows for extremely fast trajectory generation for
robots of high degrees of freedom (motion states generated within 1
millisecond). It has been used to implement robot sword fighting and
other activities that require fast reaction-based planning. The author
started a business based simply on the work and has shown the
algorithm's success in many robot applications.