Add Privacy Papers for Visual Data
We care deeply about privacy. To strengthen our knowledge in this field and understand how it relates to visual data, we do constant research in the latest scientific works about this topic. We want to share with you a curated list of machine/deep learning papers that address the topic of privacy in visual data. Although we know that it is very specific field, we think it is important to raise awareness about it.

I organised link according to this discussion https://github.com/papers-we-love/papers-we-love/issues/610