Add TDD with mutation testing (Roman & Mnich 2020)  (#1)
Add a paper on TDD+M:
Roman, A., Mnich, M. Test-driven development with mutation testing – an experimental study. Software Qual J (2020)

It presents an old idea of mutation testing in a new light of extending traditional test-driven development with it. It includes results of an experiment which was conducted to check whether TDD+M is better in terms of test efficiency than traditional TDD.